pyproject.toml
[tool.poetry]
name = "interpreter-assistant"
version = "0.1.0"
description = "Real-time bilingual transcription system"
authors = ["Mohit Sonkamble <mohitseventeens@gmail.com>"]
readme = "README.md"
packages = [{include = "interpreter_assistant"}]

[tool.poetry.dependencies]
python = "^3.10"
deepgram-sdk = "^3.2.0"
pyaudio = {version = "^0.2.14", markers = "sys_platform == 'win32'"}
websockets = "^12.0"
python-dotenv = "^1.0.0"
rich = "^13.7.0"
anyio = "^4.3.0"
numpy = "^1.26.0"

[tool.poetry.group.dev.dependencies]
black = "^23.12.0"
mypy = "^1.7.0"
pytest = "^7.4.3"
pytest-asyncio = "^0.23.5"
pytest-cov = "^4.1.0"

[tool.poetry.group.monitoring.dependencies]
prometheus-client = "^0.20.0"
psutil = "^5.9.8"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.5.3"
mkdocs-material = "^9.4.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"